Refactor QuotaRequest to ResourceRequest
