+++ troposphere/static/js/constants/ResourceRequestConstants.js
