+++ troposphere/static/js/models/ResourceRequest.js
